+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Common and useful classes, methods, exceptions etc.
 
 1. Tests > refactoring & minor improvements
 2. Utilities > CssSelector > useful methods related to CSS selectors
+3. Utilities > Bootstrap4CssSelector > useful methods related to CSS selectors and the Bootstrap4 (front-end component library)
 
 # 0.1.2